Add Methods for Calculation
Create methods to calculate the distance to spin based on current and
target colors, and include the ControlPanelAPI.

+++ FRC_2020_Infinite_Recharge_Official/src/main/java/frc/robot/APIs/ControlPanelAPI.java
@@ -137,9 +137,9 @@ public class ControlPanelAPI {
      * @param targetColorAsChar The color we are trying to spin the wheel to, as a character (can be a raw value from the FMS).
      * @return The amount of ticks to spin the motor, as a double.
      */
-    public double calculateOurSide(char targetColorAsChar) {
+    public int calculateOurSide(char targetColorAsChar) {
         double compensatedSpinDistance = 0;
-        double motorSpinDistanceInEncoders = 0; // The amount of ticks the motor should spin
+        int motorSpinDistanceInEncoders = 0; // The amount of ticks the motor should spin
         int targetColor = convertColorToInt(targetColorAsChar);        
         int currentRobotColor = calculateCurrentColor(255, 0, 0); // Add methods to get the real RGB values
         int currentWheelColor = 0;
@@ -198,9 +198,9 @@ public class ControlPanelAPI {
      * @param targetColorAsChar The color we are trying to spin the wheel to, as a character (can be a raw value from the FMS).
      * @return The amount of ticks to spin the motor, as a double.
      */
-    public double calculateEnemySide(char targetColorAsChar) {
+    public int calculateEnemySide(char targetColorAsChar) {
         double compensatedSpinDistance = 0;
-        double motorSpinDistanceInEncoders = 0; // The amount of ticks the motor should spin
+        int motorSpinDistanceInEncoders = 0; // The amount of ticks the motor should spin
         int targetColor = convertColorToInt(targetColorAsChar);        
         int currentRobotColor = calculateCurrentColor(255, 0, 0); // Add methods to get the real RGB values
         int currentWheelColor = 0;
